requirements.txt
# Core dependencies
torch>=1.9.0
numpy>=1.21.0
Pillow>=8.0.0
requests>=2.25.0

# 可选依赖（仅当使用Google API时需要）
# google-genai>=0.3.0
# google-cloud-aiplatform>=1.25.0
# google-auth>=2.0.0
# google-auth-oauthlib>=0.5.0